blow_your_mind/requirements.txt
boto3==1.24.66
botocore==1.27.66
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
Flask==2.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==9.2.0
pyaes==1.6.1
pycparser==2.21
PyJWT==2.4.0
pymongo==4.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
reportlab==3.6.11
requests==2.28.1
s3transfer==0.6.0
secure-smtplib==0.1.1
six==1.16.0
twilio==7.10.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
